Auto-PR

+++ tests/fixtures/demo/ddl_core_customer_transactions.sql
@@ -3,22 +3,20 @@
 -- Only populated for instances that have both stage_customers and stage_transactions.
 --
 -- Used in:
---   payments_test.alpha  — fully loaded (12 rows after africa filter)
---   payments_test.beta   — loaded from one business date
+--   sales_test.main — loaded with africa filter (deliberate error)
 --
 -- NOT used in:
---   sales domain         — has no transactions, so core layer is skipped
---   payments_uat         — has no transactions, so core layer is skipped
+--   payments domain — has no customers in staging, uses core_account_payments instead
 --
 -- Placeholders:
---   {database}  — e.g. payments_test
---   {schema}    — instance name, e.g. alpha
+--   {database}  — e.g. sales_test
+--   {schema}    — instance name, e.g. main
 
 CREATE SCHEMA IF NOT EXISTS {database}.{schema};
 CREATE TABLE IF NOT EXISTS {database}.{schema}.core_customer_transactions (
     id              INTEGER,
     transaction_date DATE,
-    amount          INTEGER,
+    amount          FLOAT,
     customer_id     INTEGER,
     customer_name   STRING,
     customer_type   STRING,
